refactor(gui): keep the step log change under the 700-line review bar
Move the shutdownBarrierFrom parser into runtimeClient.ts, which owns
the RPC shutdown result boundary, returning workspaceRuntime.ts to 691
lines with the tracker/barrier ownership unchanged. Split the bridge
test surface: shared harness and event factories live in
stepLogEventBridge.testSupport.ts, and the terminal/crash/supersession/
replay lifecycle cases move to stepLogEventBridge.lifecycle.test.ts —
every resulting module is under 700 lines with all assertions
preserved, plus a new guard test for render-gate replayed completions.
