+++ ecos/gui/apps/desktop-electron/electron/services/eccRpc/stepLogEventBridge.test.ts
@@ -1,95 +1,14 @@
-import { mkdtempSync, mkdirSync, readFileSync, rmSync, writeFileSync } from 'node:fs'
-import { tmpdir } from 'node:os'
-import { join } from 'node:path'
+import { rmSync } from 'node:fs'
+
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import type { EccRuntimeProtocolPayload } from '@ecos-studio/shared'
-
-import { StepLogEventBridge } from './stepLogEventBridge'
-
-function v1Marker(event: string, step: string, tool: string): Buffer {
-  return Buffer.from(
-    `\x1eECC-STEP {"v":1,"event":"${event}","step":"${step}","tool":"${tool}"}\n`,
-    'utf8',
-  )
-}
-
-function protocolEvent(
-  type: EccRuntimeProtocolPayload['type'],
-  payload: Record<string, unknown>,
-  overrides: Partial<EccRuntimeProtocolPayload> = {},
-): EccRuntimeProtocolPayload {
-  return {
-    eventId: `ecc-${type}-1`,
-    kind: 'flow',
-    operationId: 'operation-1',
-    origin: 'gui',
-    payload,
-    rerun: false,
-    runSessionId: 'run-session-1',
-    runtimeInstanceId: 'runtime-1',
-    sequence: 1,
-    timestamp: 1000,
-    type,
-    workspaceId: 'ecc-workspace-1',
-    ...overrides,
-  }
-}
-
-interface Harness {
-  bridge: StepLogEventBridge
-  workspace: string
-  emitted: EccRuntimeProtocolPayload[]
-  forwarded: EccRuntimeProtocolPayload[]
-  unscoped: string[]
-  forward: (event: EccRuntimeProtocolPayload) => void
-  feed: (...chunks: Buffer[]) => void
-  archiveText: (step: string, tool: string) => string
-}
-
-function makeHarness(
-  options: { holdTimeoutMs?: number; maxBufferedSegments?: number } = {},
-): Harness {
-  const workspace = mkdtempSync(join(tmpdir(), 'step-log-bridge-'))
-  const emitted: EccRuntimeProtocolPayload[] = []
-  const forwarded: EccRuntimeProtocolPayload[] = []
-  const unscoped: string[] = []
-  const bridge = new StepLogEventBridge({
-    workspaceDirectory: workspace,
-    emitProtocolEvent: (event) => emitted.push(event),
-    emitUnscoped: (text) => unscoped.push(text),
-    ...(options.holdTimeoutMs !== undefined
-      ? { holdTimeoutMs: options.holdTimeoutMs }
-      : {}),
-    ...(options.maxBufferedSegments !== undefined
-      ? { maxBufferedSegments: options.maxBufferedSegments }
-      : {}),
-  })
-  const forward = (event: EccRuntimeProtocolPayload) => forwarded.push(event)
-  return {
-    bridge,
-    workspace,
-    emitted,
-    forwarded,
-    unscoped,
-    forward,
-    feed: (...chunks: Buffer[]) => {
-      for (const chunk of chunks) {
-        bridge.archiver.feed(chunk)
-      }
-    },
-    archiveText: (step, tool) =>
-      readFileSync(join(workspace, `${step}_${tool}`, 'log', `${step}.log`), 'utf8'),
-  }
-}
-
-function writeFlowJson(workspace: string, steps: { name: string; tool: string }[]): void {
-  mkdirSync(join(workspace, 'home'), { recursive: true })
-  writeFileSync(
-    join(workspace, 'home', 'flow.json'),
-    JSON.stringify({ steps: steps.map((s) => ({ ...s, state: 'Unstart' })) }),
-  )
-}
+import {
+  makeHarness,
+  protocolEvent,
+  v1Marker,
+  writeFlowJson,
+  type Harness,
+} from './stepLogEventBridge.testSupport'
 
 describe('StepLogEventBridge', () => {
   let workspaceDirs: string[] = []
@@ -498,265 +417,5 @@ describe('StepLogEventBridge', () => {
     h.feed(v1Marker('begin', 'Synthesis', 'yosys'), Buffer.from('after replay\n'))
     h.feed(v1Marker('end', 'Synthesis', 'yosys'))
     expect(h.emitted[0]!.sequence).toBe(4)
-  })
-
-  it('resets lifecycle state on sidecar close', () => {
-    const h = harness()
-    writeFlowJson(h.workspace, [{ name: 'Synthesis', tool: 'yosys' }])
-    startOperation(h, 42)
-    startStep(h, 'Synthesis', 'yosys', 43)
-    h.bridge.handleSidecarClose()
-
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'operation.started',
-        {},
-        {
-          operationId: 'operation-2',
-          runSessionId: 'run-session-2',
-          sequence: 1,
-        },
-      ),
-      h.forward,
-    )
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'step.started',
-        { step: 'Synthesis', tool: 'yosys', state: 'Ongoing' },
-        { operationId: 'operation-2', runSessionId: 'run-session-2', sequence: 2 },
-      ),
-      h.forward,
-    )
-    h.feed(v1Marker('begin', 'Synthesis', 'yosys'), Buffer.from('after restart\n'))
-    h.feed(v1Marker('end', 'Synthesis', 'yosys'))
-    expect(h.emitted[0]!.sequence).toBe(2)
-    expect(h.emitted[0]!.operationId).toBe('operation-2')
-  })
-
-  it('holds a terminal operation event behind a held step.completed', () => {
-    const h = harness()
-    writeFlowJson(h.workspace, [{ name: 'Synthesis', tool: 'yosys' }])
-    startOperation(h)
-    startStep(h, 'Synthesis', 'yosys')
-    h.feed(v1Marker('begin', 'Synthesis', 'yosys'), Buffer.from('failing output\n'))
-
-    // A failed step does not wait for the render gate: step.completed and
-    // operation.failed can both arrive before the stderr end marker.
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'step.completed',
-        { step: 'Synthesis', tool: 'yosys', state: 'Imcomplete' },
-        { sequence: 3 },
-      ),
-      h.forward,
-    )
-    h.bridge.handleProtocolEvent(
-      protocolEvent('operation.failed', { error: { message: 'x' } }, { sequence: 4 }),
-      h.forward,
-    )
-    expect(h.forwarded.map((event) => event.type)).toEqual([
-      'operation.started',
-      'step.started',
-    ])
-
-    // The end marker finally arrives: completion first, then the terminal.
-    h.feed(v1Marker('end', 'Synthesis', 'yosys'))
-    expect(h.forwarded.map((event) => event.type)).toEqual([
-      'operation.started',
-      'step.started',
-      'step.completed',
-      'operation.failed',
-    ])
-    expect(h.forwarded[2]!.payload.finalLog).toBe('failing output\n')
-  })
-
-  it('forwards a queued terminal on hold timeout after the completion', () => {
-    vi.useFakeTimers()
-    const h = harness({ holdTimeoutMs: 2000 })
-    writeFlowJson(h.workspace, [{ name: 'Synthesis', tool: 'yosys' }])
-    startOperation(h)
-    startStep(h, 'Synthesis', 'yosys')
-    h.feed(v1Marker('begin', 'Synthesis', 'yosys'), Buffer.from('some output\n'))
-
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'step.completed',
-        { step: 'Synthesis', tool: 'yosys', state: 'Imcomplete' },
-        { sequence: 3 },
-      ),
-      h.forward,
-    )
-    h.bridge.handleProtocolEvent(
-      protocolEvent('operation.failed', { error: { message: 'x' } }, { sequence: 4 }),
-      h.forward,
-    )
-
-    vi.advanceTimersByTime(2000)
-    expect(h.forwarded.map((event) => event.type)).toEqual([
-      'operation.started',
-      'step.started',
-      'step.completed',
-      'operation.failed',
-    ])
-  })
-
-  it('abandons the stale archive on timeout so the next run starts fresh', () => {
-    vi.useFakeTimers()
-    const h = harness({ holdTimeoutMs: 2000 })
-    writeFlowJson(h.workspace, [{ name: 'Synthesis', tool: 'yosys' }])
-    startOperation(h)
-    startStep(h, 'Synthesis', 'yosys')
-    h.feed(v1Marker('begin', 'Synthesis', 'yosys'), Buffer.from('first attempt\n'))
-
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'step.completed',
-        { step: 'Synthesis', tool: 'yosys', state: 'Imcomplete' },
-        { sequence: 3 },
-      ),
-      h.forward,
-    )
-    vi.advanceTimersByTime(2000)
-    expect(h.forwarded.map((event) => event.type)).toContain('step.completed')
-
-    // The next attempt's begin must not read as nested inside the stale one.
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'operation.started',
-        {},
-        {
-          operationId: 'operation-2',
-          runSessionId: 'run-session-2',
-          sequence: 4,
-        },
-      ),
-      h.forward,
-    )
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'step.started',
-        { step: 'Synthesis', tool: 'yosys', state: 'Ongoing' },
-        { operationId: 'operation-2', runSessionId: 'run-session-2', sequence: 5 },
-      ),
-      h.forward,
-    )
-    h.feed(v1Marker('begin', 'Synthesis', 'yosys'), Buffer.from('second attempt\n'))
-    h.feed(v1Marker('end', 'Synthesis', 'yosys'))
-
-    expect(h.archiveText('Synthesis', 'yosys')).toBe('second attempt\n')
-    const synthesized = h.emitted.filter((event) => event.type === 'step.log')
-    expect(synthesized.at(-1)?.operationId).toBe('operation-2')
-    expect(synthesized.at(-1)?.payload.chunk).toBe('second attempt\n')
-  })
-
-  it('tracks attempts across reruns when completion precedes the end marker', () => {
-    const h = harness()
-    writeFlowJson(h.workspace, [{ name: 'Synthesis', tool: 'yosys' }])
-    startOperation(h)
-    startStep(h, 'Synthesis', 'yosys')
-
-    // First attempt: completion arrives before its end marker.
-    h.feed(v1Marker('begin', 'Synthesis', 'yosys'), Buffer.from('attempt one\n'))
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'step.completed',
-        { step: 'Synthesis', tool: 'yosys', state: 'Success' },
-        { sequence: 3 },
-      ),
-      h.forward,
-    )
-    expect(h.forwarded.map((event) => event.type)).not.toContain('step.completed')
-    h.feed(v1Marker('end', 'Synthesis', 'yosys'))
-    expect(h.forwarded[2]!.payload.finalLog).toBe('attempt one\n')
-
-    // Rerun the same step: its completion must also wait for its own end.
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'step.started',
-        { step: 'Synthesis', tool: 'yosys', state: 'Ongoing' },
-        { sequence: 4 },
-      ),
-      h.forward,
-    )
-    h.feed(v1Marker('begin', 'Synthesis', 'yosys'), Buffer.from('attempt two\n'))
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'step.completed',
-        { step: 'Synthesis', tool: 'yosys', state: 'Success' },
-        { sequence: 5 },
-      ),
-      h.forward,
-    )
-    const completedBefore = h.forwarded.filter((event) => event.type === 'step.completed')
-    expect(completedBefore).toHaveLength(1)
-    h.feed(v1Marker('end', 'Synthesis', 'yosys'))
-    const completedAfter = h.forwarded.filter((event) => event.type === 'step.completed')
-    expect(completedAfter).toHaveLength(2)
-    expect(completedAfter[1]!.payload.finalLog).toBe('attempt two\n')
-  })
-
-  it('resolves a held completion and queued terminal before a new operation starts', () => {
-    const h = harness()
-    writeFlowJson(h.workspace, [{ name: 'Synthesis', tool: 'yosys' }])
-    startOperation(h)
-    startStep(h, 'Synthesis', 'yosys')
-    h.feed(v1Marker('begin', 'Synthesis', 'yosys'), Buffer.from('failed output\n'))
-
-    // The step's completion is held (no end marker yet) and the terminal
-    // event queues behind it.
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'step.completed',
-        { step: 'Synthesis', tool: 'yosys', state: 'Imcomplete' },
-        { sequence: 3 },
-      ),
-      h.forward,
-    )
-    h.bridge.handleProtocolEvent(
-      protocolEvent('operation.failed', { error: { message: 'x' } }, { sequence: 4 }),
-      h.forward,
-    )
-    expect(h.forwarded.map((event) => event.type)).toEqual([
-      'operation.started',
-      'step.started',
-    ])
-
-    // ecc allows a new operation once the previous one is terminal: its
-    // start must first resolve the leftover hold and terminal in order.
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'operation.started',
-        {},
-        {
-          operationId: 'operation-2',
-          runSessionId: 'run-session-2',
-          sequence: 5,
-        },
-      ),
-      h.forward,
-    )
-    expect(h.forwarded.map((event) => event.type)).toEqual([
-      'operation.started',
-      'step.started',
-      'step.completed',
-      'operation.failed',
-      'operation.started',
-    ])
-    expect(h.forwarded[2]!.payload.finalLog).toBe('failed output\n')
-    expect(h.forwarded[4]!.operationId).toBe('operation-2')
-
-    // The new operation's markers start a fresh archive attempt.
-    h.bridge.handleProtocolEvent(
-      protocolEvent(
-        'step.started',
-        { step: 'Synthesis', tool: 'yosys', state: 'Ongoing' },
-        { operationId: 'operation-2', runSessionId: 'run-session-2', sequence: 6 },
-      ),
-      h.forward,
-    )
-    h.feed(v1Marker('begin', 'Synthesis', 'yosys'), Buffer.from('fresh run\n'))
-    h.feed(v1Marker('end', 'Synthesis', 'yosys'))
-    expect(h.archiveText('Synthesis', 'yosys')).toBe('fresh run\n')
-    expect(h.emitted.at(-1)?.operationId).toBe('operation-2')
   })
 })
